Move dataset structure to script
General tidying up of code
