+++ models/data.go
@@ -1,18 +1,5 @@
 package models
 
-// Datasets ...
-type Datasets struct {
-	Items []Dataset `json:"items"`
-}
-
-// Dataset ...
-type Dataset struct {
-	Alias       string `json:"alias"`
-	Description string `json:"description"`
-	Link        string `json:"link"`
-	Title       string `json:"title"`
-}
-
 type SearchResponse struct {
 	Hits Hits `json:"hits"`
 }
